+++ src/app/services/powershell/powershell-commands.ts
@@ -50,7 +50,7 @@ export class PowershellCommands {
         if (computerName) command += ` -ComputerName "${computerName}"`;
         command += ' -List';
 
-        AppLogger.getInstance().logDebug(command);
+        AppLogger.getDebug().logDebug(command);
         return PsCommandExecutor.executeCommand(command, false, 90)
         .then(output => PowershellCommands._parseEventViewersList(output))
         .then(events => {
@@ -60,7 +60,7 @@ export class PowershellCommands {
     }
     
     private static _parseEventViewersList(output: string): EventLog[] {
-        AppLogger.getInstance().logDebug(output);
+        AppLogger.getDebug().logDebug(output);
         const regex = /\S*\s+\d+\s+\D+\s+([\d|,|.]+)\s+(.*)/s;
 
         return GlobalUtils.splitLines(output)   //get each line